split types in `staking` module
