+++ x/staking/src/types/dv.rs
@@ -1,0 +1,29 @@
+use gears::types::address::{AccAddress, ValAddress};
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct DvvTriplet {
+    pub del_addr: AccAddress,
+    pub val_src_addr: ValAddress,
+    pub val_dst_addr: ValAddress,
+}
+impl DvvTriplet {
+    pub fn new(del_addr: AccAddress, val_src_addr: ValAddress, val_dst_addr: ValAddress) -> Self {
+        Self {
+            del_addr,
+            val_src_addr,
+            val_dst_addr,
+        }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct DvPair {
+    pub val_addr: ValAddress,
+    pub del_addr: AccAddress,
+}
+impl DvPair {
+    pub fn new(val_addr: ValAddress, del_addr: AccAddress) -> Self {
+        Self { val_addr, del_addr }
+    }
+}
