+++ x/staking/src/types/mod.rs
@@ -1,8 +1,14 @@
-mod domain;
+mod delegations;
+mod dv;
+mod historical_info;
 pub mod iter;
 mod query;
-mod tx_proto;
+mod tx;
+mod validator;
 
-pub use domain::*;
+pub use delegations::*;
+pub use dv::*;
+pub use historical_info::*;
 pub use query::*;
-pub use tx_proto::*;
+pub use tx::*;
+pub use validator::*;
